refactor: AppLanguage enum을 common으로 이동
