+++ app/src/main/java/com/eatssu/android/presentation/mypage/language/LanguageSelectorViewModel.kt
@@ -5,7 +5,7 @@ import androidx.core.os.LocaleListCompat
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.eatssu.android.data.local.SettingDataStore
-import com.eatssu.android.domain.model.AppLanguage
+import com.eatssu.common.enums.AppLanguage
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
