+++ core/common/src/main/java/com/eatssu/common/enums/AppLanguage.kt
@@ -1,4 +1,4 @@
-package com.eatssu.android.domain.model
+package com.eatssu.common.enums
 
 import java.util.Locale
